Security Fix for the Define Cog
Corrections per issue #241.

Removes all non-number and letter characters besides &, _, and -. This may have unintended consquences interfering with words using accents or something like that.

As a second line of defense, exceptions are evaluated for whether they will report the API key. If so, the API key is removed from the exception's error message.